DRT-4761 Removed code that is no longer used.
:)
