+++ client/src/main/scala/drt/client/components/Debug.scala
@@ -23,9 +23,6 @@ object Debug {
           m.shiftsRaw,
           m.fixedPointsRaw,
           m.staffMovements,
-          m.workloadPot,
-          m.queueCrunchResults,
-          m.simulationResult,
           m.crunchStatePot,
           m.loadingState
         )
@@ -34,22 +31,16 @@ object Debug {
         Pot[String],
           Pot[String],
           Pot[Seq[StaffMovement]],
-          Pot[Workloads],
-          RootModel.PortCrunchResults,
-          RootModel.PortSimulationResults,
           Pot[CrunchState],
           LoadingState
         )]) => {
-        val (potShifts, potFixedPoints, staffMovements, potWorkloads, tqcr, simulationResult, crunchState, loadingState) = staffingMP()
+        val (potShifts, potFixedPoints, staffMovements, crunchState, loadingState) = staffingMP()
 
         if (dom.window.hasOwnProperty("debug")) {
           <.table(
             <.tr(<.th("shifts"), potShifts.renderReady(s => <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), s)))),
             <.tr(<.th("fixed points"), potFixedPoints.renderReady(s => <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), s)))),
             <.tr(<.th("staff movements"), <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), staffMovements.toString()))),
-            <.tr(<.th("Terminal queue Crunch Results"), <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), tqcr.toString()))),
-            <.tr(<.th("Simulation Result"), <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), simulationResult.toString()))),
-            <.tr(<.th("workloads"), potWorkloads.renderReady(s => <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), s.toString)))),
             <.tr(<.th("crunch State"), crunchState.renderReady(s => <.td(<.pre(^.style := js.Dictionary("overflow" -> "auto", "height" -> "200px"), s.toString)))),
             <.tr(<.th("loading state"), s"$loadingState")
           )
