+++ client/src/main/scala/drt/client/components/FlightsTable.scala
@@ -1,15 +1,12 @@
 package drt.client.components
 
-//import chandu0101.scalajs.react.components.Spinner
 import diode.data.{Pot, Ready}
-import drt.shared.{ArrivalHelper, PaxTypesAndQueues}
-//import japgolly.scalajs.react.vdom.all.{VdomAttr => _, TagMod => _, _react_attrString => _, _react_autoRender => _, _react_fragReactNode => _}
-import japgolly.scalajs.react.vdom.html_<^._
+import drt.shared.PaxTypesAndQueues
 import drt.client.modules.{FlightsView, GriddleComponentWrapper, ViewTools}
 import drt.shared.AirportInfo
-import drt.shared.FlightsApi.{Flights}
+import drt.shared.FlightsApi.Flights
 import japgolly.scalajs.react._
-
+import japgolly.scalajs.react.vdom.html_<^._
 
 import scala.language.existentials
 import scala.scalajs.js
@@ -65,7 +62,7 @@ object FlightsTable {
         props.flightsWithSplitsPot.renderPending((t) => ViewTools.spinner),
         props.flightsWithSplitsPot.renderEmpty(ViewTools.spinner),
         props.flightsWithSplitsPot.renderReady(flights => {
-            FlightsWithSplitsTable.ArrivalsTable()()(FlightsWithSplitsTable.Props(flights, props.bestPax, PaxTypesAndQueues.inOrderSansFastTrack.toList))
+            FlightsWithSplitsTable.ArrivalsTable()()(FlightsWithSplitsTable.Props(flights, props.bestPax, PaxTypesAndQueues.inOrderSansFastTrack))
         })
       )
     }).build
