+++ client/src/main/scala/drt/client/components/TerminalComponent.scala
@@ -2,26 +2,19 @@ package drt.client.components
 
 import diode.data.{Pending, Pot}
 import diode.react.ModelProxy
-import drt.client.actions.Actions.{HideLoader, SetPointInTime, ShowLoader}
 import drt.client.components.FlightComponents.SplitsGraph.splitsGraphComponentColoured
 import drt.client.components.FlightComponents.paxComp
-import drt.client.components.Heatmap.Series
-import drt.client.components.TerminalHeatmaps._
 import drt.client.logger.log
-import drt.client.services.HandyStuff.QueueStaffDeployments
 import drt.client.services.JSDateConversions.SDate
-import drt.client.services.{SPACircuit, TimeRangeHours, Workloads}
+import drt.client.services.{SPACircuit, TimeRangeHours}
 import drt.shared.Crunch.CrunchState
-import drt.shared.FlightsApi.{FlightsWithSplits, QueueName, TerminalName}
-import drt.shared.Simulations.QueueSimulationResult
+import drt.shared.FlightsApi.TerminalName
 import drt.shared._
 import japgolly.scalajs.react.extra.Reusability
 import japgolly.scalajs.react.vdom.html_<^
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.{BackendScope, Callback, ScalaComponent}
-import org.scalajs.dom
 
-import scala.collection.immutable.Map
 import scala.util.Try
 
 object TerminalComponent {
@@ -32,11 +25,6 @@ object TerminalComponent {
                             crunchStatePot: Pot[CrunchState],
                             airportConfig: Pot[AirportConfig],
                             airportInfos: Pot[AirportInfo],
-                            simulationResult: Map[QueueName, QueueSimulationResult],
-                            crunchResult: Map[QueueName, CrunchResult],
-                            deployments: QueueStaffDeployments,
-                            workloads: Workloads,
-                            actualDesks: Map[QueueName, Map[Long, DeskStat]],
                             pointInTime: Option[SDateLike],
                             timeRangeHours: TimeRangeHours
                           )
@@ -46,11 +34,6 @@ object TerminalComponent {
       model.crunchStatePot,
       model.airportConfig,
       model.airportInfos.getOrElse(props.terminalName, Pending()),
-      model.simulationResult.getOrElse(props.terminalName, Map()),
-      model.queueCrunchResults.getOrElse(props.terminalName, Map()),
-      model.staffDeploymentsByTerminalAndQueue.getOrElse(props.terminalName, Map()),
-      model.workloadPot.getOrElse(Workloads(Map())),
-      model.actualDeskStats.getOrElse(props.terminalName, Map()),
       model.pointInTime,
       model.timeRangeFilter
     ))
@@ -64,11 +47,6 @@ object TerminalComponent {
             airportConfig,
             props.terminalName,
             model.airportInfos,
-            model.simulationResult,
-            model.crunchResult,
-            model.deployments,
-            model.workloads,
-            model.actualDesks,
             model.timeRangeHours,
             model.pointInTime.getOrElse(SDate.today())
           )
@@ -86,72 +64,13 @@ object TerminalComponent {
 
   val component = ScalaComponent.builder[Props]("Terminal")
     .renderPS(($, props, state) => render(props))
-      .componentDidUpdate(p => Callback.log("Updating Terminal Component"))
-      .componentDidMount(p => Callback.log("Updating Terminal Component"))
+    .componentDidUpdate(p => Callback.log("Updating Terminal Component"))
+    .componentDidMount(p => Callback.log("Updating Terminal Component"))
     .build
 
   def apply(props: Props): VdomElement = {
     component(props)
   }
-}
-
-
-object HeatmapComponent {
-
-  case class Props(
-                    airportConfig: AirportConfig,
-                    terminalName: TerminalName,
-                    simulationResults: Map[QueueName, QueueSimulationResult]
-                  ) {
-    lazy val hash = simulationResults.values.map(_.hashCode).hashCode()
-  }
-
-  case class State(activeTab: String)
-
-  implicit val propsReuse = Reusability.by((_: Props).hash)
-  implicit val stateReuse = Reusability.caseClass[State]
-
-  val component = ScalaComponent.builder[Props]("Heatmaps")
-    .initialState(State("deskrecs"))
-    .renderPS((scope, props, state) =>
-      <.div({
-        val seriesPot: Pot[List[Series]] = waitTimes(props.simulationResults, props.terminalName)
-        val baseHeight = 120
-        val heatmapQueuesHeight = props.airportConfig.queues(props.terminalName).length * 40
-        val heatmapsContainerHeight = baseHeight + heatmapQueuesHeight
-        <.div(^.height := s"${heatmapsContainerHeight}px",
-          <.ul(^.className := "nav nav-tabs",
-            <.li(^.className := "active", <.a(VdomAttr("data-toggle") := "tab", ^.href := "#deskrecs", "Desk recommendations"), ^.onClick --> scope.modState(_ => State("deskrecs"))),
-            <.li(<.a(VdomAttr("data-toggle") := "tab", ^.href := "#workloads", "Workloads"), ^.onClick --> scope.modState(_ => State("workloads"))),
-            <.li(<.a(VdomAttr("data-toggle") := "tab", ^.href := "#paxloads", "Paxloads"), ^.onClick --> scope.modState(_ => State("paxloads"))),
-            <.li(seriesPot.renderReady(s => {
-              <.a(VdomAttr("data-toggle") := "tab", ^.href := "#waits", "Wait times", ^.onClick --> scope.modState(_ => State("waits")))
-            }))
-          )
-          ,
-          <.div(^.className := "tab-content",
-            <.div(^.id := "deskrecs", ^.className := "tab-pane fade in active",
-              if (state.activeTab == "deskrecs") {
-                heatmapOfStaffDeploymentDeskRecs(props.terminalName)
-              } else ""),
-            <.div(^.id := "workloads", ^.className := "tab-pane fade",
-              if (state.activeTab == "workloads") {
-                heatmapOfWorkloads(props.terminalName)
-              } else ""),
-            <.div(^.id := "paxloads", ^.className := "tab-pane fade",
-              if (state.activeTab == "paxloads") {
-                heatmapOfPaxloads(props.terminalName)
-              } else ""),
-            <.div(^.id := "waits", ^.className := "tab-pane fade",
-              if (state.activeTab == "waits") {
-                heatmapOfWaittimes(props.terminalName, props.simulationResults)
-              } else "")
-          ))
-      }))
-    .configure(Reusability.shouldComponentUpdate)
-    .build
-
-  def apply(props: Props): VdomElement = component(props)
 }
 
 object TerminalContentComponent {
@@ -161,22 +80,13 @@ object TerminalContentComponent {
                     airportConfig: AirportConfig,
                     terminalName: TerminalName,
                     airportInfoPot: Pot[AirportInfo],
-                    simulationResult: Map[QueueName, QueueSimulationResult],
-                    crunchResult: Map[QueueName, CrunchResult],
-                    deployments: QueueStaffDeployments,
-                    workloads: Workloads,
-                    actualDesks: Map[QueueName, Map[Long, DeskStat]],
                     timeRangeHours: TimeRangeHours,
                     dayToDisplay: SDateLike
                   ) {
     lazy val hash = {
-      val depsHash: List[Option[List[Int]]] = deployments.values.map(drtsPot => {
-        drtsPot.toOption.map(drts => {
-          drts.items.map(drt => {
-            drt.hashCode
-          }).toList
-        })
-      }).toList
+      val depsHash = crunchStatePot.map(
+        cs => cs.crunchMinutes.toSeq.map(_.hashCode())
+      ).toList.mkString("|")
 
       val flightsHash: Option[List[(Int, String, String, String, String, String, String, String, String, Long, Int)]] = crunchStatePot.toOption.map(_.flights.toList.map(f => {
         (f.splits.hashCode,
@@ -242,7 +152,6 @@ object TerminalContentComponent {
       <.div(
         <.ul(^.className := "nav nav-tabs",
           <.li(^.className := "active", <.a(VdomAttr("data-toggle") := "tab", ^.href := "#arrivals", "Arrivals"), ^.onClick --> t.modState(_ => State("arrivals"))),
-          <.li(<.a(VdomAttr("data-toggle") := "tab", ^.href := "#queues", "Desks & Queues Old"), ^.onClick --> t.modState(_ => State("queues"))),
           <.li(<.a(VdomAttr("data-toggle") := "tab", ^.href := "#desksAndQueues", "Desks & Queues"), ^.onClick --> t.modState(_ => State("desksAndQueues"))),
           <.li(<.a(VdomAttr("data-toggle") := "tab", ^.href := "#staffing", "Staffing"), ^.onClick --> t.modState(_ => State("staffing")))
         ),
@@ -259,25 +168,9 @@ object TerminalContentComponent {
               }))
             } else ""
           }),
-          <.div(^.id := "queues", ^.className := "tab-pane fade terminal-desk-recs-container",
-            if (state.activeTab == "queues") {
-              val deploymentProps = TerminalDeploymentsTable.TerminalProps(
-                props.airportConfig,
-                props.terminalName,
-                props.crunchStatePot,
-                props.simulationResult,
-                props.crunchResult,
-                props.deployments,
-                props.workloads,
-                props.actualDesks,
-                props.timeRangeHours
-              )
-              TerminalDeploymentsTable.terminalDeploymentsComponent(deploymentProps)
-            } else ""
-          ),
           <.div(^.id := "desksAndQueues", ^.className := "tab-pane fade terminal-desk-recs-container",
             if (state.activeTab == "desksAndQueues") {
-              props.crunchStatePot.renderReady( crunchState => {
+              props.crunchStatePot.renderReady(crunchState => {
 
                 TerminalDesksAndQueues(TerminalDesksAndQueues.Props(crunchState, props.airportConfig, props.terminalName))
               })
