--- client/src/test/scala/drt/client/services/StaffQueueAssignmentTests.scala
@@ -1,11 +1,0 @@
-package drt.client.services
-
-import utest._
-
-object StaffQueueAssignmentTests extends TestSuite {
-  def tests = TestSuite {
-    'DeskRecHandler - {
-
-    }
-  }
-}
